API of `subscribe`: Takes a message constructor function instead of a mailbox address
This way the user has the chance to transform the response before posting it into the mailbox.
This is in line with the conventions used in the standard library (e.g. `Html.Events.on`).

Nonetheless, we may switch from mailbox-based reporting to task-based reporting.
So, this commit is primarily to document the right way to do mailbox-based reporting, even if we don't keep it.
It may reappear as a convenience function later.