API of `subscribe`: Report cancellations by running a separate task.
The first parameter of `subscribe` is a function used to construct a task from a data response.
The second parameter is a function used to construct a task from a cancellation.

We split that case out because in many apps you will likely do nothing on a cancellation, or at least something totally different compared to data reception.

If you don't need to respond to cancellations, the following may be a handy convenience function for the second parameter:

    doNothing : a -> Task x ()
    doNothing = always (Task.succeed ())